+++ docs/CHANGELOG.md
@@ -1,5 +1,7 @@
 ## Next
 
+* Assign and display mods to an armour slot they will fit with in LO
+
 ## 6.9.0 <span className="changelog-date">(2020-05-03)</span>
 
 * In the Loadout Optimizer, mods have been split into their own menu, separate from perks.